[Rewards] Remove unused notification types
