+++ components/brave_rewards/resources/rewards_page/components/notifications/external_wallet_disconnected.tsx
@@ -1,16 +1,18 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
+/* Copyright (c) 2025 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 import * as React from 'react'
 
-import { getExternalWalletProviderName } from '../../lib/external_wallet'
-import { LocaleContext, formatMessage } from '../../lib/locale_context'
-import { ExternalWalletDisconnectedNotification } from './notification'
+import { useLocaleContext } from '../../lib/locale_strings'
+import { getExternalWalletProviderName } from '../../../shared/lib/external_wallet'
+import { formatMessage } from '../../../shared/lib/locale_context'
+import { ExternalWalletDisconnectedNotification } from '../../lib/app_state'
 import { NotificationViewProps } from './notification_view'
 
 export function ExternalWalletDisconnected (props: NotificationViewProps) {
-  const { getString } = React.useContext(LocaleContext)
+  const { getString } = useLocaleContext()
   const { Title, Body, Action } = props
   const { provider } = props.notification as ExternalWalletDisconnectedNotification
