+++ components/brave_rewards/resources/rewards_page/components/notifications/notification_view.ts
@@ -1,10 +1,11 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
+/* Copyright (c) 2025 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 import * as React from 'react'
 
-import { Notification, NotificationAction } from './notification'
+import { Notification, NotificationAction } from '../../lib/app_state'
 
 export interface NotificationTitleProps {
   style?: 'funding' | 'information' | 'error' | 'custom'
