+++ components/brave_rewards/resources/rewards_page/lib/app_state.ts
@@ -4,7 +4,6 @@
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 import { ExternalWallet, ExternalWalletProvider } from '../../shared/lib/external_wallet'
-import { Notification } from '../../shared/components/notifications'
 import { ProviderPayoutStatus } from '../../shared/lib/provider_payout_status'
 import { Optional } from '../../shared/lib/optional'
 
@@ -122,7 +121,33 @@ export interface CaptchaInfo {
   maxAttemptsExceeded: boolean
 }
 
-export { Notification }
+export type NotificationType =
+  'monthly-tip-completed' |
+  'external-wallet-disconnected'
+
+export interface Notification {
+  type: NotificationType
+  id: string
+  timeStamp: number
+}
+
+export interface ExternalWalletDisconnectedNotification extends Notification {
+  type: 'external-wallet-disconnected'
+  provider: ExternalWalletProvider
+}
+
+export type NotificationActionType =
+  'open-link' |
+  'reconnect-external-wallet'
+
+export interface NotificationAction {
+  type: NotificationActionType
+}
+
+export interface OpenLinkNotificationAction extends NotificationAction {
+  type: 'open-link'
+  url: string
+}
 
 export { ExternalWallet, ExternalWalletProvider }
